Heavy! Here comes the sources of gdx-setup-ui v3
Full commit for the src and libs folders. A loooooot of things changed
since last revision. Most noticable one is the support for updating an
existing project with the latest version of libgdx and any third-party
lib.

See https://github.com/AurelienRibon/gdx-setup-ui for the list of all
the in-between commits.