getRotation for livewallpaper
But it's little bit buggy, don't know why.
If you place your device to 0 degrees position (portraite orientation) 
and rotate to 180 degrees, nothing be detected, or from 90 to 270 and 
versa. But if you rotate your device from 0 to 90 or 270, or from 90 
to 180 and 0 and so on - all is fine (different orientations).