Prevent duplicate mouse up events from ruining mouse input.
This was reproduced by double clicking a xfiledialog. The second click causes it to disappear and GLFW gives us a mouse up with no corresponding mouse down. Maybe because it is a modal child of the GLFW window? Hope there is no way to get incorrect mouse downs, or we'll be screwed.