Add new tests for TiledMapAtlas direct loading and AssetManager loading
Add resources (+preprocessed) to be able to run tests on checkout.

Note about resources: the "tiled-atlas-src" directory contains the source
used to build the "tiled-atlas-processed" resources: you can build the
preprocessor by running "ant -f build.xml gdx-tiled-preprocessor" in the
root libgdx/ directory.

Once built, run it against the source data, specifying an input and an
output directory:

<packer> <libgdx/.../tiled-atlas-src> <libgdx/.../tiled-atlas-processed>

Note that at this time the preprocessor will NOT overwrite the destination
folder if it already contains the same data, so make sure it's clean.