List:  added ability to select 'nothing', and ability to disable item selection (+ bug/defect fix)
This adds functionality to the List Actor so it may behave like that of a ListView akin to Android's.  This is backwards compatible.

This is essential in situations where you may not always want to have an item selected (like a touch-based file browser), or where you may not want items even selectable at all (like during a tutorial).

The bug/defect fix was if you look in the old setSelection(String name) method, you'll see it sets the selectedIndex to -1, then tries to find an item with that name.  When it doesn't, it leaves the selectedIndex at -1 which essentially sets it to nothing.


*Note*
When the List is NOT selectable, it will no longer eat its touch event meaning they will still propagate down.

I'm not 100% sure if that is the behavior you want versus the contrary where it eats the touch event despite not doing anything.  Regardless, I have a note of this behavior change in the setSelectable comment.