Made Payload available on dragstop without target
There are some situations in which a drag starts with a payload and the payload is removed from the source (e. g. Inventory). When the dragging does NOT stop over a target, then only `dragStop` is called on the source, but the payload is not supplied anymore and that makes it more difficult to put the payload back to the source.

The additional payload parameter makes this situation much easier to handle.