Fixes ShapeRenderer multicolor call to rect()
So, it seems that if you called rect() using all the parameters
(position, size, origin, rotation and 4 colors for the vertices),
if the ShapeRenderer was open on Filled mode, it actually ignored
any colors you gave it and render the rect on the current renderer
color (the one you set by calling setColor).

I fixed it by making the function use the colors you give it,
just like the other functions that take colors as parameters do.