Add support for converting popular Android ARGB8888 format
As requested, replaces PR #2410 with a clean change history.

In a recent code review, I found two versions of this function in my code, so I thought it might be a useful addition. My code originally was Android-only before porting to libGDX, so perhaps others will be in this same situation and will find it convenient.