This changes permit to customize the textbutton labels.
Skin.get now also returns resource that are subclasses of the given
type.
Implemented a method createLabel in the Label.LabelStyle class, which
permits to instantiate the proper subclass of Label without knowing it.
Added a labelStyle field to TextButton.TextButtonStyle which allows
TextButton to use a subclass of Label instantiating it with the
createLabel method.

Added DynamicChildActorsTest to test the above functionalities.