Moved BitmapFont setScale and containsCharacter to BitmapFontData.
I left the getLineHieght and other getters in BitmapFont since they are convenient and are used a lot. It doesn't make sense to have setters (like setScale) in BitmapFont though. Modification of the BitmapFontData should be done through BitmapFontData, else it makes it seem like the setting is per BitmapFont.