Fix: non-white labels have glitchy color upon size calc.
If a tinted label (having its color set via setColor instead of style
font) was not wrapped or had ellipses set then the label’s layout’s
font color would not be set to white but would instead be set to it’s
last used color. Therefore when draw() tinted the layout you ended up
with compounded and incorrect tinting.