Read kerning data from the GPOS table in addition to the kern table
This also cleans up the Kerning class a bit, adds a ByteInputStream
subclass to handle data decoding, reads all of the file into memory in
one go because we need to seek (and it's probably faster), and uses more
efficient data structures (IntIntMap instead of lots of
boxing/unboxing).

Fixes #3442.