Adjust Stage mouse input
Adjusts mouse input y-coordinate to be relative to the viewport's height instead of the window's height.  This was a necessary change to my own game where I render a stage to a framebuffer at a constant screen and world size, so as to prevent distortion from texture filtering on individual actors due to my assets being extremely low resolution.  Since there are times when my window might be a larger height than the framebuffer's resolution, and consequently the stage's resolution, the input would be incorrectly positioned and handled if not for this change.