Increased fidelity of progress tracking.
Added a simple counter of total tasks created by individual asset loads,
which can be compared against the size of the current task stack to
determine fractional progress towards completion of a single asset.
This increases the fidelity of the progress value returned, which is
especially useful for assets with a large number of dependencies.