Allow tap count reset on ClickListener
I was having a hard time working around this missing method. There are cases where I have a widget that does something different with a drag than with a double click. But if I quickly drag and then click soon after the drag, it still counts as a double click.