Added the setInputProcessor.
Added the setInputProcessor. 
In the current class and the parent class, was not set neither in the input processor.

I do not know if the ApplyForce class is extended in some other example and that other example there is the set for inputprocessor. If ApplyForce really is a complete example then missed this set.