Added support for typed TMX-properties.
Modified the loadProperties() method by adding support for detecting the
types of the different properties. The type-information is stored as an
additional attribute with the name "type" in the property-element. If
present, this type-information is then used to convert the value-string
to the correct type (i.e. Boolean, Color, Float, FileHandle or Integer).