Add kotlin source for kotlin project instead of java source files in gdx-setup
When users select to generate project with kotlin language, then most
likely those users want to completely work with kotlin. So generated
project should reflect this too, thus this changes offer that.

All launcher files are in .kt file now whenever "Use Kotlin" is selected in Advanced section of gdx-setup.

Also note about HTML platform target as the code check gwtSupported flag
of Language whether it supports or not, and it's not hard-coded. Thus it
offers flexibility for devs in the future to just change the flag when
gwt could build HTML target working with kotlin.