Refactor specific resource for Java and Kotlin out
Specific resource used for either Java or Kotlin language is refactored
out in _specifics/ folder in which inside there are java/ and kotlin/
folder. Inside of each language folder, intended to keep the same
structure as of generated project, thus src/ is there that contain the
same name of launcher class files.

GdxSetup.java contains an additional of private method to return out
such resource path of src according to language, and project type.