Allow List#setItems() with the original items
This allows uses like the following, without the need for copying the
entire array:

    Array<String> items = list.getItems();
    items.add("hello world");
    list.setItems(items);

Issue #2834 suggests that this should work, but it doesn't; it clears
the items instead.

It seems unlikely that anyone would use that on purpose (as
`List#clearItems()` already exists), so this change is unlikely to break
anyone.