Add Poolable#onObtain and Poolable#onFree
Created default methods onObtain() and onFree() within the Poolable interface so that they may be used for state changes that a user may want. For example, the use case that gave me this idea was trying to check if an object was in active or in a pool and there was currently no plain & simple way to do it.

It's been implemented into Pool#freeAll() , Pool#free() , and Pool#obtain() and will work with this implementation due to the nature of Pools' methods using the Pool#free/all/obtain methods as opposed to their own implementation.

Any comments, concerns, suggestions, etcetera would be appreciated-- and I will be signing the agreement shortly.