make ParticleController deltaTime not hard-coded (#5030)
* -The ParticleController stores a frame time in its deltaTime member but this
value is hard-coded at 1/60 regardless of framerate. This change adds an overload to the
ParticleController's update() method to take the delta time. This allows the particles to adjust
their time deltas for faster or flower framerate. The change also
allows callers to cause particle effects to be played back at faster or slower
speeds.
-This change also adds overloads for update() and updateAndDraw() of the ParticleSystem class,
each of which takes a frame time and adds an overload for the ParticleEffect update() method.
-RegularEmitter previously only used the floor of the number of milliseconds that had passed since
 the last update when calculating percent of the lifetime lived for each particle. For playback of
 particle effects that were being played at .1 times the speed of the default speed, .00016666 ms
were elapsed between calls to update(), which rounds down to 1 ms rather than 1.666 ms.
This meant that the particle effect being played at .1 speed was actually being animated for 16 times as long rather than 10 times as long as the same effect played back at 1x speed
(because the floor of .001666 *1000 = 16 but floor of .00016666*1000 is 1.)
-This change also updates the Flame tool to make it respect the delta multiplier

* Forgot to add the FlameMain.java changes