Small update to Intersector intersectSegmentPlane
On rare occasions, Intersector.intersectSegmentPlane would set the intersect vector to (NaN,NaN,NaN) and return true. I found this was due to divide by zero error from denom. This could be my problem of not validating parameters but it would be helpful if a check was added to return false if denom == 0 or within epsilon.