[freetype] Include shadow size in cap height.
The AngleCode font format stores only "base", which is capHeight + ascent. The capHeight is not stored, instead it is gotten from a glyph height. A bitmap font's glyph height includes shadow, so the FreeType font's capHeight must also. The achieved goal is identical rendering of a FreeType font and that same font turned into a bitmap font using BitmapFontWriter.