[scene2d] TextField should not rely on focus events to determine focused state.
Also reverts "[scene2d] When a focus event is cancelled, fire another event to put it back.", commit 837456aa2d816d003a73a5b0293269f50e096666. We don't fire a second event for any other event type when the change is cancelled, so it's odd to do it only for keyboard focus.