Fixing touchpad scroll under OSX
OSX generates many events when scrolling with the touch pad. Lwjgl3
reports small movement with a scroll value lower than 1. This is
rounded to 1 or -1. If there are many such "small" events the
scrolling becomes ultra fast.

The current fix acculates the scrolling reported by Lwjgl3 and sends a
Gdx scroll event when the accumulated value is greater than 1. This is
done until the accumulated value is lower than 1. This approach also
supports accelerations which are reported as scrolling value greater
than 1.

The drawback of the current approach is that it does not send a scroll
event until the accumulated value is greater than 1. Doing this would
need to use a float in the gdx event and change the API.