Added debug mode to I18NBundle (#5427)
When developing a project, it's common to hardcode strings during testing rather than localizing them. Usually the developer will miss lots of unlocalized strings. A common method to spot unlocalized strings is to create another localization file with some placeholder value like "X" for every single key. This makes it easy to see what strings haven't been localized in game. To avoid having to maitain another localization file, the debug() method in I18NBundle will set all strings to placeholder to simplify this process. 

Since I18NBundle.format() can take parameters but ignores them if it can't find a place to put them in the string, the method should work fine. Also, the placeholder can be set to anything, so it can always be set to "{0}, {1}, {2}..." to capture parameters. Although this shouldn't be an issue if they are ignored as expected.