Improved LWJGL2 backend pause/resume.
Previously pause happened when the window was no longer the foreground window, but rendering continued. This created strange/surprising behavior, such as Timer not firing (because the app is paused) but rendering was still happening.

Now pause doesn't happen by default when not the foreground window. It can be enabled using `pauseWhenBackground`, which doesn't stop rendering (use `backgroundFPS=-1` to stop rendering). It's still weird to pause and continue rendering, but some apps might rely on this to know when they are in the background, eg to pause gameplay but continue rendering to show a dialog.

Previously minimizing/restoring the window did not pause/resume. Now it does by default.