Enhanced Animation.java.
Spares the user from keeping track of the stateTime, which is also not safe. Instead, it keeps track of its own stateTime, just by calling its update(float deltaTime) method, for every frame, opposed to the previous approach, where the user had to keep track and update the stateTime, on their own. 
Ofcourse, if the user wants to keep track of one or multiple stateTime, they can, just by calling the previous methods.

A few other methods were added, to manipulate the stateTime.