Added notEmpty() convenience methods.
It is nice to express the intent clearly with this convenience method. It would be best to express it positively, but there's no good, generic method name for the opposite of empty. `something.notEmpty()` is at least better than `!something.isEmpty()`, where the not and `isEmpty` are separated. `isEmpty` and `notEmpty` are pretty similar, but the difference is still bigger than not having or having a `!`.

#5199