[scene2d.ui] Removed @Override.
It makes sense for application code. It isn't required for library code, where it just adds noise and the benefits for refactoring are insignificant.

I removed it only from the classes I authored/maintain. Removing the thousands in other code would be nice, but I don't care as much and others like the clutter.