Ensure BitmapFontCache.setColors(float, int, int) stays in-bounds
This only needs a single `Math.min(int, int)` call to be added per call to `setColors(float, int, int)`, checking to make sure that the calculated end index is no greater than what other parts of the class use (`idx[page]`, where page is always 0 in the broken case). I added the reproduction code from libgdx/libgdx#5834 to BitmapFontTest, which needed an if/else block to be changed to actually run the tests on most of the features.