Fix two subtle issues, write a test for them.
First, all of the maps and sets didn't check adequately for constructors with initial capacity 0; thanks to @code-disaster, this was found and the fix simply initializes the backing array to at minimum 2 items. This was a one-line fix per file. Second, and more insidious, resizing wasn't always correctly changing the size field of int- and long-keyed maps and sets because 0 keys are handled specially there; the fix is another line changed in each.