GlyphLayout: moved calculating run widths and the entire layout's width to after the runs are determined.
This simplifies things a tiny bit, since we don't calculate the run width as we go, we don't have to adjust it when the run glyphs are wrapped or truncated. Instead we have to iterate all the runs and glyphs at the end. This gives us a chance to detect when a glyph is wider than the right edge of subsequent glyphs, which closes #6098.