GlyphLayout, moved pools and color stack to statics.
The class was not thread safe before this commit, as it uses pools. Discussion long ago in #3491. At least now it is clear in the javadoc.

The statics should be fine with Android resumes, as the Pools class is.

Also changed in this commit is that the color pool isn't cleaned up until the next time it is used. This doesn't appear to be an issue and makes it a little nicer to do all the markupEnabled work in one place.