Fixed GestureDetector maxFlingDelay.
In old code `time - tracker.lastTime` was always 0. Maybe this never worked? Plus the javadocs for the parameter did not match the name of the parameter nor what the code appeared to be trying to do.

I also increased GestureDetector maxFlingDelay from 0.15 to Integer.MAX_VALUE. With 0.15 flings rarely happen. With a more reasonable default like 0.5, existing code would still be in for a surprise: when the parameter starts working, fling behavior would change. With Integer.MAX_VALUE app code relying on the defaults doesn't see a difference.

I'm not really a fan of this setting anyway, as it seems better to always fling (Integer.MAX_VALUE) and let app code discard flings where the touch down is too long ago, the velocity is too low, or whatever else. But we have this setting, so it should at least work!