AndroidControllers: possibility to not ignore some buttons on a controller (#5702)
ffe15f0d introduced that AndroidControllers extension ignore every button a controller has that is not considered a Gamepad button. However, some users could be irritated that some buttons on their controller can't be used. With this commit, the developer can switch to the old behaviour and the game will get events for all buttons the controller provides.