Fix to #4489 , requesting review re: sorting and/or geometry
I'm not as certain why this works as I'd like to be; if anyone wants to investigate why the extra check is needed for the sort to be correct, I would appreciate it. The repro case in #4489 should clearly show that without this PR, `points` isn't sorted correctly (index 10 should be the highest x position, but it isn't), and with this PR it is sorted correctly. I also don't think I can run the tests while there's an Ant build script trying to invoke `/bin/bash` and I'm on Windows.