Fix mistake in recently-added put() for ObjectFloatMap and ObjectLongMap (#6186)
* Specify encoding for Base64Coder.encodeString(), as the rest of GDX does

This commit also adds Javadocs where they were missing for one encodeString() and makes the spelling of "Base64" consistently "Base64" and not sometimes "Base 64".

* Add line to CHANGES for #6061

* Fix to #4489 , requesting review re: sorting and/or geometry

I'm not as certain why this works as I'd like to be; if anyone wants to investigate why the extra check is needed for the sort to be correct, I would appreciate it. The repro case in #4489 should clearly show that without this PR, `points` isn't sorted correctly (index 10 should be the highest x position, but it isn't), and with this PR it is sorted correctly. I also don't think I can run the tests while there's an Ant build script trying to invoke `/bin/bash` and I'm on Windows.

* Simple fix for put() typos on ObjectFloatMap, ObjectLongMap

The types were wrong, probably due to copy/paste, for the value parameter in both places.

* Avoid wildcard import of all of ObjectSet; applies to maps and sets

This is an addendum; it's an automatically-added import change that appears more correct. I don't think the maps and sets should static-import everything in ObjectSet, since they only use tableSize(). I was under the impression that libGDX tried to avoid wildcard imports, though I don't know why.