Make Bresenham2 actually implement Bresenham's algorithm
Before, it implemented something else, and couldn't complete some common lines, like any 2-point diagonal line. The test is now much more robust, and can be used to preview any line from the center of a 160x120 area to anywhere in that area. If green is ever visible in the test's line, that means there's some discrepancy between the actual algorithm and what Bresenham2 implements; so far I haven't seen any green and this can connect all lines. This fixes #5012 .