[scene2d.ui] Table#round uses ceil/floor and is applied during layout. Table clean up.
Previously layout was done without rounding, then the final child positions/sizes rounded. This meant a child can be smaller/larger than when computing the layout -- eg it could overlap another cell by 1 or be 1 too small. This is especially noticeable when a label that uses ellipsis truncation is 1px too small (the label is truncated when it should have fit).

Ceil is better for sizes because it's preferable the actor is slightly larger than too small (eg a label with ellipsis again). Floor is better for position because if something is slightly larger and also happens to round up then it's even farther off the desired layout.