NinePatch, fixed blurry rendering when using a single center region.
new NinePatch(TextureRegion) places a single region in the center without surrounding regions. This is useful eg to specify a background that doesn't stretch but does have padding (eg a slider background). Before this commit the 0.5 texel UV shift was always applied to the center, even if there are no surrounding regions, resulting in blurry rendering. Fixed by not applying the UV shift unless adjacent patches have a region.