Update the Gradle version gdx-setup uses to 6.7
This allows gdx-setup to actually complete its generation on Java 13, 14, and 15. The wrapper files are also updated. No changes appear to be needed to any build.gradle files for desktop or android, but  I can't try ios to see if Gradle 6.7 is compatible. The html platform works when using libGDX 1.9.12-SNAPSHOT, but of course fails on the current stable version for the same reasons this failed before.