Fix getIndices() crash in Mesh (#6390)
Fix getIndices() crash in Mesh, credit to Agueliethun

This was a copy/paste error from the Buffer casts change recently. The getIndices() code mistakenly changed the vertices buffer where it formerly changed only the indices buffer. This caused crashes in ModelCacheTest when the Cache checkbox was selected and any model was added.