Hiero: Default to crash-less Java renderer, return to it if Freetype would cause crash (#6403)
* If a font lacks 'x', fall back to Arial and place a note.

The note can be easily removed just by pressing the ASCII, NeHe, or Extended buttons, but it lets the user know why Arial was selected instead of some emoji-only font or something.

* If a font is missing 'x' and we use Freetype, switch to Java.

Only Freetype rendering has an issue with missing 'x' chars. Java and Native don't necessarily show anything if the font has no ASCII chars, but they don't crash.

* We don't need the 'x' check if we default to Java rendering.

This also enables the effects right away, which I think are a major reason people use Hiero.

Also, now the Java radio button is selected safely if 'x' is missing and Freetype is currently selected. The bold and italic checkboxes are also restored as part of that.