Clean up MathUtils.random() for generating longs (#6412)
* Complex fix for MathUtils.random(long) .

This one's an especially intricate piece of code, and the bug it fixes is sorta non-obvious. Before, if you called `MathUtils.nextLong(Long.MAX_VALUE)` many times, you would never get a number with a `1` bit in any of the bottom 10 bits. This is obviously not random; the lack of odd numbers might be the most easily noticeable. That happens because the old code used a double (with 52 bits of mantissa) to generate a random long (with 64 bits). The ensuing precision loss affects all large-enough ranges. The fix is to use Rafael Baptista's technique for getting the upper 64 bits of a 64x64-to-128-bit multiply, treating the inputs as unsigned. It still works for negative bounds, like before, and is inclusive. It also has a speedup if the bound is one less than a power of two (such as for `Long.MAX_VALUE`), since not as much work is needed there.

* Fix up random(long, long) so larger ranges can be produced.

* Allow UUIDs to use the sign bit.

Before, they were only positive longs for each portion of the UUID, and before this PR, they were only using 53 bits of each 64-bit long.

* Add a test to make sure.

Co-authored-by: Nathan Sweet <nathan.sweet@gmail.com>