Improve precision for MathUtils.sinDeg(180f)
This particular right angle was given special treatment by MathUtils, but was still wrong, yielding -8.742278E-8 instead of 0. I think the updated code is actually less complex, because it doesn't risk a difference between the double-precision radians that Math.sin() expects and the single-precision radians MathUtils can provide it. I also added tests for sinDeg() and cosDeg(), which would have failed before because of the extremely low tolerance they use.