Remove old stuff and add new screenshot to gdx-setup README.md
Removed the "files altered" list because it's very incomplete. I tried updating it but it turns out gdx-setup modifies a lot of files. I was struggling to ensure the list is complete and questioned what purpose it would actually serve, as it's safe to assume everything in `resources/` is at risk of getting altered in one way or another.

Updated the image. This screenshot is taken on Windows, as gdx-setup looks a bit naff on Linux. I don't really use Windows but hopefully the filepaths and whatnot look legitimate.

Completely removed the assets information, as its location should now be obvious.

Some other minor changes included - capital G for Gradle, etc.