Fix "nitpicks," private atn -> public atanUnchecked().
There's some documentation improvements for atanUnchecked().
I also removed local final modifiers, and else before return, for all of the inverse trig methods (since this had snuck into my earlier asin() and acos() code). The docs for atan() point out atanUnchecked() as an option, and that it only offers a tiny speedup, if that.