Split natives fetching, fix publish workflow
The GitHub publish workflow builds libGDX natives on the fly and uses the upload/download artifact actions to get the natives before deployment to SonaType. Some new extensions depend on natives not build by the libGDX build, but provided in external properties. Since the publish workflow doesn't invoke the fetchNatives task, these external natives are not part of the build. This commit fixes this by splitting up the natives fetching tasks and invoking the external native fetcher task during publishing.