[scene2d.ui] WidgetGroup#hit first validates the layout.
Actors can be in the wrong position if the layout is not valid before hit(). For example, if an actor was added on the previous frame and layout has not yet occurred. In that case the actor can be hit (usually by an `enter` input event), then just before being drawn the layout is validated and the actor is drawn at a position where it would not have been hit. Buttons and other actors using the `enter` event (or hit) for a hover state are drawn in the hover state for 1 frame, even though they were never hovered.