Reuse the OS checks from SharedLibraryLoader.
This UIUtils class isn't used on GWT, so SharedLibraryLoader should be safe to use in all cases. I've built the tests to target GWT, Android, LWJGL, and LWJGL3, but not iOS because I don't have a Mac. On the four platforms I built the tests on, the build succeeded; I ran the tests without issue on LWJGL 2 and 3.

I should note that GWT failed to build once due to some bug internal to Spotless (related to OSGi?), but built fine on the next attempt. That might explain some random failures in automated tests...