For GL30 compatibility, use indexed lsans-15.png .
I don't know the specifics for why GL30 has a hard time with the Grayscale+Alpha color mode for PNGs. But, it doesn't seem to load that correctly, and does load Indexed color mode PNGs just fine, so switching to a smaller indexed-mode file should allow the default font to work on all GL versions. There's a tiny bit of lost detail, but this takes advantage of PNG's ability to have different alpha levels for every color in a palette, so there should still be a soft, not jagged, edge.