Trying to track down other grayscale+alpha PNGs...
This should be most or all of the problem PNGs in the tests, but I might have missed some. I used TinyPNG to do a lossy compression to 256 colors (including alpha in each color entry), and this doesn't look bad to me so far. The images probably didn't have more than 256 colors to begin with.