Fixed some typos in Vector-related docs.
This also changes several methods to use Java-style array declaration, when internally they had been using C-style array declaration. Using the C-style arrays is a style warning in IDEA and probably could be in Spotless too.

I wanted to clear up that idt() means "identical to" here, and not "identity" as it does in Quaternion, so I added one line of extra doc to each idt() in Vector3 and Vector4. I also mentioned that it compares with exact precision, not like MathUtils.isEqual() .