Do not force-overwrite the LWJGL configuration option GLFW_LIBRARY_NAME
Right now, GLFW_LIBRARY_NAME *always* gets overwritten. That's an issue for people using this property, as there is no way to get around this. The proposal is to only rewrite the configuration if no custom location was used, i.e. the property still has the default value of `null`.