Do not inline the version string
Few will be aware of the constant inlining done by javac so this may
hurt some library authors (though outside of that few should be affected
by the constant inlining). Plainly appending .toString() makes so
javac does not inline the string, meaning that the version string
can be used to obtain the runtime version of libGDX, not the version
used to compile the library with.

Note: Obviously this won't retroactively affect older binaries,
so older libraries that are compiled with older versions of libGDX
may still display this anomaly