Added tiledmap loader for JSON files. Replacement for PR (#7386) (#7505)
* Rebasing against latest map related commit - 98d64e2

* Fixed issue where the BaseTmjMapLoader was not properly loading groups.
Fixed imagelayers inside group not being loaded in TmjMapLoader

* Fixed issue where objects were not loading because of an unneeded if statement in loadObjects().
Fixed an issue causing a crash when loading a String property, we now check for "string" as a type in the castProperty() method.

* Added <include name="maps/tiled/BaseTmjMapLoader.java"/> and <include name="maps/tiled/TmjMapLoader.java"/> to the gdx.gwt.xml file.

* Added a new Test, TiledMapJsonFormatTest.
This test is basically a copy of the TiledMapLayerOffsetTest. But I converted the .tmx files for the maps to .tmj json format. Also converted the .tsj file for the tilesets used in the ortho maps.
Added 1 extra orthographic map at the end which contains nested groups and tile animations. Just to make sure it's supported.

* Apply formatter

* Fixed a crash in .tmj maps when loading tilesets which are just a collection of images.

* Small Refactor across the Base MapLoader classes. BaseTmjMapLoader and BaseTmxMapLoader now both extend BaseTiledMapLoader which is where all common shared methods will live.

Added new test and testmap for loading objects from .tmj maps

Removed loadCustomTileSet() methods as they were unrelated to the PR and not function as requested.

* Apply formatter

* Renamed a Tileset used in a test File.

---------

Co-authored-by: GitHub Action <action@github.com>